pyproject.toml
[tool.poetry]
name = "srvmon"
version = "0.1.0"
description = "A simple yet useful server status monitoring tool."
authors = ["Harry Cheng <chengyuhui1@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
flask = "^1.1.2"
flask-sqlalchemy = "^2.4.4"
flask-socketio = "^4.3.2"
APScheduler = "^3.6.3"
python-dotenv = "^0.15.0"
Flask-Migrate = "^2.5.3"
email_validator = "^1.1.2"
flask-security-too = "^3.4.4"
eventlet = "^0.29.1"
Flask-Mail = "^0.9.1"
Flask-WTF = "^0.14.3"
multidict = "^5.1.0"
cerberus = "^1.3.2"
flask-restx = "^0.2.0"
requests = "^2.25.1"
blinker = "^1.4"
pytest = "^6.2.1"

[tool.poetry.dev-dependencies]
yapf = "^0.30.0"
mock = "^4.0.3"
requests-mock = "^1.8.0"

[[tool.poetry.source]]
name = "tuna"
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = true

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
